requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asttokens==3.0.0
async-generator==1.10
attrs==25.3.0
binaryornot==0.4.4
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cim-graph==0.1.7a0
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
cryptography==37.0.4
dataclasses-json==0.5.14
dateutils==0.6.12
defusedxml==0.7.1
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
Flask==2.3.3
Flask-Session==0.5.0
flask-talisman==1.1.0
flet==0.22.1
flet-core==0.22.1
flet-runtime==0.22.1
gevent==21.12.0
greenlet==1.1.3.post0
grequests==0.6.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
icecream==2.1.4
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
neo4j==5.28.1
nest-asyncio==1.6.0
oauthlib==3.2.2
outcome==1.3.0.post0
oxrdflib==0.3.7
packaging==23.2
pickleDB==0.9.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymysql==1.1.1
pyOpenSSL==22.0.0
pyoxigraph==0.3.22
pyparsing==3.2.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
qrcode==7.4.2
rdflib==7.1.3
rdflib-neo4j==0.0.1b6
repath==0.9.0
requests==2.28.2
rich==13.9.4
simplekv==0.14.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.40
starlette==0.46.1
stomp.py==6.0.0
text-unidecode==1.3
trio==0.21.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==4.3.1
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
xsdata==24.12
zope.event==5.0
zope.interface==7.2
